pyproject.toml
[project]
name = "pixv"
version = "0.1.2"
description = "Local Version Control System for Images"
authors = [
  { name = "Som Shekhar Sharma", email = "sharmathat@gmail.com" },
  { name = "Dhruv Sagar", email = "dhruvsagar24nd@gmail.com"}
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "pillow>=10.0.1",
  "numpy>=1.26.1",
  "requests>=2.31.0",
  "matplotlib>=3.8.2",
  "click>=8.1.7",
]
requires-python = ">=3.10.0"

[tool.hatchling.dependencies]
python = ">=3.10.0"
pillow = ">=10.0.1"
numpy = ">=1.26.1"
requests = ">=2.31.0"
matplotlib = ">=3.8.2"
click = ">=8.1.7"

[project.scripts]
pixv = "pixv.feature:cli"

# [tool.package_name]
# data_folder = "src/Graphs"


[tool.hatch.build.targets.wheel]
packages = ["pixv"]


[project.urls]
"Homepage" = "https://github.com/ReallyJaegar/PixV"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

